[#809] Rewrote the routing middleware + tests
Tests needs to be cleaned up, probably to different test cases per
accessed host. Updated Piwik template tag to reflect changes.
