+++ akvo/rsr/context_processors.py
@@ -53,8 +53,8 @@ def extra_pages_context(request):
             'return_url': page.return_url,
             'return_url_text': page.custom_return_url_text,
             'stylesheet': page.stylesheet,
-            'akvoapp_root_url': request.akvoapp_root_url,
-            'domain_url': request.domain_url,
+            'akvoapp_root_url': '//{}'.format(settings.AKVOAPP_DOMAIN),
+            'domain_url': '//{}'.format(settings.RSR_DOMAIN),
             'no_facebook': not page.facebook_button,
             'facebook_app_id': page.facebook_app_id,
             'no_twitter': not page.twitter_button,
