+++ akvo/settings/10-base.conf
@@ -61,8 +61,8 @@ AUTHENTICATION_BACKENDS = (
 AUTH_USER_MODEL = 'rsr.User'
 
 MIDDLEWARE_CLASSES = (
-    'akvo.rsr.middleware.PagesRouterMiddleware',
-    # 'akvo.rsr.middleware.PagesLocaleMiddleware',
+    'akvo.rsr.middleware.HostDispatchMiddleware',
+    # 'akvo.rsr.middleware_old.PagesLocaleMiddleware',
     # 'django.middleware.gzip.GZipMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
