+++ akvo/settings/66_local.template
@@ -5,7 +5,7 @@
     Akvo RSR module. For additional details on the GNU license please
     see < http://www.gnu.org/licenses/agpl.html >.
 """
-
+SITE_ID = 1
 DEBUG = TEMPLATE_DEBUG = True
 PIPELINE_ENABLED = False
