+++ akvo/templates/base.html
@@ -102,7 +102,6 @@
         }
     </script>
 
-    {# Piwik #}
-    {% tracking_code %}
+    {% piwik_tracking_code %}
   </body>
 </html>
